pyproject.toml
[project]
name = "laneige-ranking-agent"
version = "0.1.0"
description = "LANEIGE Global Ranking Insight Agent"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv==1.0.0",
    "anthropic==0.40.0",
    "pandas==2.1.4",
    "openpyxl==3.1.2",
    "chromadb==0.4.22",
    "sentence-transformers==2.2.2",
    "fastapi==0.109.0",
    "uvicorn==0.27.0",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.25",
    "rich==13.7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pandas-stubs>=2.1.4",
    "types-openpyxl>=3.1.0",
]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["backend"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["backend"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "venv",
    "__pycache__",
]

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "sentence_transformers.*",
]
ignore_missing_imports = true
